agrocaixa-starter/services/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic-settings==2.10.1
SQLAlchemy
psycopg2-binary==2.9.11
alembic
